refactor: delegate product goal views
